+++ tests/functional/test_cross_entropy.py
@@ -6,8 +6,10 @@ import torch
 import torch.distributed as dist
 import torch.multiprocessing as mp
 
-from fast_llm.functional.config import CrossEntropyImpl, TargetFormat
+from fast_llm.functional.config import CrossEntropyImpl, TargetFormat, TritonConfig
 from fast_llm.functional.cross_entropy import cross_entropy_forward_backward, reverse_kl_forward_backward
+from fast_llm.utils import Assert
+from tests.utils.utils import requires_cuda
 
 
 def _mp_worker(rank: int, world_size: int, init_method: str, fn_args: tuple):
@@ -175,10 +177,76 @@ def combined_worker(rank: int, group: dist.ProcessGroup, use_mask: bool):
     _ce_vocab_tp_worker(rank, group, use_mask)
 
 
+# TODO: maybe merge these tests using same parametrization
 @pytest.mark.slow
 @pytest.mark.parametrize("use_mask", [True, False])
-def test_reverse_combined(use_mask):
+def test_distillation_losses(use_mask):
     _spawn_dist(2, combined_worker, use_mask)
+
+
+@requires_cuda
+@pytest.mark.slow
+@pytest.mark.parametrize(
+    ("num_columns", "grad_output", "logits_scale_factor", "loss_masking"),
+    (
+        (8192, 1.0, 1.0, False),  # Simple
+        (5000, 1.0, 1.0, False),  # Not a power of 2
+        (5000, None, 1.0, False),  # No grad
+        (5000, 1.0, 4.0, False),  # Loss scaling
+        (5000, 4.0, 1.0, False),  # Grad scaling
+        (5000, 1.0, 1.0, True),  # Loss masking
+        (65536, 1.0, 1.0, False),  # Max block size
+        (65537, 1.0, 1.0, False),  # Above max block size
+    ),
+)
+@pytest.mark.parametrize("target_format", (TargetFormat.labels, TargetFormat.logits, TargetFormat.probabilities))
+def test_cross_entropy(num_columns, grad_output, logits_scale_factor, loss_masking, target_format):
+    # TODO: Test tensor-parallel implementation.
+    assert TritonConfig.TRITON_ENABLED
+    # We want something moderately close to the target for the test to be meaningful
+    logits_var = torch.randn(256, num_columns, dtype=torch.bfloat16, device="cuda") / 3
+    loss_mask = torch.randint(0, 2, (256,), dtype=torch.bool, device="cuda") if loss_masking else None
+    if target_format == TargetFormat.labels:
+        target = torch.randint(0, num_columns, (256,), dtype=torch.int64, device="cuda")
+        logits = (torch.nn.functional.one_hot(target, num_columns) + logits_var).requires_grad_()
+        if loss_masking:
+            logits = torch.where(loss_mask.unsqueeze(-1), logits, -100)
+            loss_mask = None
+    else:
+        target = torch.randn(256, num_columns, dtype=torch.bfloat16, device="cuda")
+        logits = (target + logits_var).requires_grad_()
+        if target_format == TargetFormat.probabilities:
+            target = torch.softmax(target, -1)
+
+    kwargs = {
+        "logits": logits,
+        "target": target,
+        "loss_mask": loss_mask,
+        "grad_output": grad_output,
+        "logits_scale_factor": logits_scale_factor,
+        "target_format": target_format,
+    }
+    # Torch serves as the reference implementation.
+    out_torch, grad_torch = cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.torch)
+
+    out_fused, grad_fused = cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.fused)
+    Assert.rms_close(out_fused, out_torch, 5e-3)
+    if grad_output is None:
+        assert grad_torch is None
+        assert grad_fused is None
+    else:
+        Assert.rms_close(grad_fused, grad_torch, 5e-3)
+
+    if num_columns > 65536:
+        with pytest.raises(AssertionError):
+            cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.triton)
+    else:
+        out_triton, grad_triton = cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.triton)
+        if grad_output is None:
+            assert grad_triton is None
+        else:
+            Assert.rms_close(grad_triton, grad_torch, 5e-3)
+        Assert.rms_close(out_triton, out_torch, 5e-3)
 
 
 if __name__ == "__main__":
