+++ tests/functional/test_triton_kernels.py
@@ -2,14 +2,7 @@ import pytest
 import torch
 
 from fast_llm.engine.config_utils.tensor_dim import TensorDim
-from fast_llm.functional.config import (
-    MAX_DROPLESS_BLOCK_SIZE_ROW,
-    ActivationType,
-    CrossEntropyImpl,
-    TargetFormat,
-    TritonConfig,
-)
-from fast_llm.functional.cross_entropy import cross_entropy_forward_backward
+from fast_llm.functional.config import MAX_DROPLESS_BLOCK_SIZE_ROW, ActivationType, TritonConfig
 from fast_llm.functional.triton.adam import triton_adam
 from fast_llm.functional.triton.mlp import (
     torch_mlp_activation,
@@ -192,71 +185,6 @@ def test_triton_mlp_activation(gated, activation, recompute):
     Assert.rms_close(input_grad1, input_.grad, 1e-5)
     if recompute:
         Assert.rms_close(output1, output3, 1e-5)
-
-
-@requires_cuda
-@pytest.mark.slow
-@pytest.mark.parametrize(
-    ("num_columns", "grad_output", "logits_scale_factor", "loss_masking"),
-    (
-        (8192, 1.0, 1.0, False),  # Simple
-        (5000, 1.0, 1.0, False),  # Not a power of 2
-        (5000, None, 1.0, False),  # No grad
-        (5000, 1.0, 4.0, False),  # Loss scaling
-        (5000, 4.0, 1.0, False),  # Grad scaling
-        (5000, 1.0, 1.0, True),  # Loss masking
-        (65536, 1.0, 1.0, False),  # Max block size
-        (65537, 1.0, 1.0, False),  # Above max block size
-    ),
-)
-@pytest.mark.parametrize("target_format", (TargetFormat.labels, TargetFormat.logits, TargetFormat.probabilities))
-def test_cross_entropy(num_columns, grad_output, logits_scale_factor, loss_masking, target_format):
-    # TODO: Test tensor-parallel implementation.
-    assert TritonConfig.TRITON_ENABLED
-    # We want something moderately close to the target for the test to be meaningful
-    logits_var = torch.randn(256, num_columns, dtype=torch.bfloat16, device="cuda") / 3
-    loss_mask = torch.randint(0, 2, (256,), dtype=torch.bool, device="cuda") if loss_masking else None
-    if target_format == TargetFormat.labels:
-        target = torch.randint(0, num_columns, (256,), dtype=torch.int64, device="cuda")
-        logits = (torch.nn.functional.one_hot(target, num_columns) + logits_var).requires_grad_()
-        if loss_masking:
-            logits = torch.where(loss_mask.unsqueeze(-1), logits, -100)
-            loss_mask = None
-    else:
-        target = torch.randn(256, num_columns, dtype=torch.bfloat16, device="cuda")
-        logits = (target + logits_var).requires_grad_()
-        if target_format == TargetFormat.probabilities:
-            target = torch.softmax(target, -1)
-
-    kwargs = {
-        "logits": logits,
-        "target": target,
-        "loss_mask": loss_mask,
-        "grad_output": grad_output,
-        "logits_scale_factor": logits_scale_factor,
-        "target_format": target_format,
-    }
-    # Torch serves as the reference implementation.
-    out_torch, grad_torch = cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.torch)
-
-    out_fused, grad_fused = cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.fused)
-    Assert.rms_close(out_fused, out_torch, 5e-3)
-    if grad_output is None:
-        assert grad_torch is None
-        assert grad_fused is None
-    else:
-        Assert.rms_close(grad_fused, grad_torch, 5e-3)
-
-    if num_columns > 65536:
-        with pytest.raises(AssertionError):
-            cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.triton)
-    else:
-        out_triton, grad_triton = cross_entropy_forward_backward(**kwargs, implementation=CrossEntropyImpl.triton)
-        if grad_output is None:
-            assert grad_triton is None
-        else:
-            Assert.rms_close(grad_triton, grad_torch, 5e-3)
-        Assert.rms_close(out_triton, out_torch, 5e-3)
 
 
 @requires_cuda
